# Conflicts:
#   .docker_local/start_scripts/build/ros1.bash
#   .docker_local/start_scripts/build/ros2.bash

+++ .docker_local/start_scripts/build/ros1.bash
@@ -2,7 +2,7 @@
 
 DONE_FILE=install/.done
 cd "${HOME}"/march/ros1/ || exit
-rm "${DONE_FILE}"
+if [ -f "${DONE_FILE}" ]; then rm "${DONE_FILE}"; fi;
 source /opt/ros/noetic/setup.bash
 colcon build --cmake-args -DCMAKE_EXPORT_COMPILE_COMMANDS=ON
-exec touch "${DONE_FILE}"
+exec touch ${DONE_FILE}
